+++ crates/fsci-conformance/tests/diff_stats_hypergeom.rs
@@ -222,7 +222,7 @@ fn diff_stats_hypergeom() {
         let oracle = pmap.get(&case.case_id).expect("validated oracle");
         let dist = Hypergeometric::new(case.big_m, case.n, case.big_n);
         if let Some(spmf) = oracle.pmf {
-            let d = (dist.pmf(case.k) - spmf).abs();
+            let d = (dist.pmf(case.k as i64) - spmf).abs();
             max_overall = max_overall.max(d);
             diffs.push(CaseDiff {
                 case_id: case.case_id.clone(),
@@ -232,7 +232,7 @@ fn diff_stats_hypergeom() {
             });
         }
         if let Some(scdf) = oracle.cdf {
-            let d = (dist.cdf(case.k) - scdf).abs();
+            let d = (dist.cdf(case.k as i64) - scdf).abs();
             max_overall = max_overall.max(d);
             diffs.push(CaseDiff {
                 case_id: case.case_id.clone(),
