+++ crates/fsci-conformance/tests/golden_stats_zipfian.rs
@@ -53,8 +53,8 @@ fn golden_stats_zipfian() {
 
     for case in &table.points {
         let dist = Zipfian::new(case.a, case.n);
-        let pmf_diff = (dist.pmf(case.k) - case.pmf).abs();
-        let cdf_diff = (dist.cdf(case.k) - case.cdf).abs();
+        let pmf_diff = (dist.pmf(case.k as i64) - case.pmf).abs();
+        let cdf_diff = (dist.cdf(case.k as i64) - case.cdf).abs();
         max_pmf = max_pmf.max(pmf_diff);
         max_cdf = max_cdf.max(cdf_diff);
         assert!(
